added a larger test for the stackmachine and parser
This creates an interesting conundrum. I wanted to implement a
haphazard debugger with some breakpoints.
I think we need to change the behaviour that instructions are
calling the next instructions.
The other thing which is not ideal is, that the main Run method
always runs both evaluation and execution is a fixed mode.
Another thing we need to change is to create a Gene type and
have a program be comprised of n genes.